pyproject.toml
[project]
name = "sl-api"
version = "0.1.0"
description = "Text classification and form completion API using BAML + FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "baml-py>=0.75.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.0.0",
]